cloud/application/requirements.txt
tensorflow==2.14.0
Flask==3.0.0
boto3==1.29.6
python-dotenv==1.0.0
Pillow==10.1.0